requirements.txt
alabaster==0.7.9
argon2-cffi==16.2.0
astroid==1.4.8
Babel==2.3.4
bcrypt==3.1.0
cffi==1.7.0
click==6.6
coverage==4.2
Django==1.10
django-appconf==1.0.2
django-compressor==2.1
django-debug-toolbar==1.5
docutils==0.12
ecdsa==0.13
Fabric==1.12.0
first==2.0.1
flake8==3.0.4
imagesize==0.7.1
isort==4.2.5
Jinja2==2.8
lazy-object-proxy==1.2.2
MarkupSafe==0.23
mccabe==0.5.2
paramiko==1.17.2
pep8==1.7.0
pip-tools==1.7.0
psycopg2==2.6.2
pycparser==2.14
pycrypto==2.6.1
Pygments==2.1.3
pylint==1.6.4
pytest==3.0.3
pytz==2016.6.1
rcssmin==1.0.6
rjsmin==1.0.12
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.4.6
sqlparse==0.2.1
tox==2.4.1
wrapt==1.10.8

djangorestframework==3.5.1
django-webpack-loader==0.3.3
jsonpickle==0.9.3
django-user-agents==0.3.0
django-extensions==1.7.4
